requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

folium==0.15.1
google-cloud-aiplatform==1.37.0
langchain==0.0.337
llama-index==0.9.13
numpy==1.25.2
pandas==2.1.3
Pillow==10.0.0
python-dotenv==1.0.0
replicate==0.21.1
requests==2.31.0
streamlit==1.29.0
streamlit-folium==0.17.2
tenacity==8.2.3
tqdm==4.66.1
trulens-eval==0.18.3
gcloud
llamahub

google_oauth2_tool
pyowm
